final code

+++ Firmware/src/children/sensors/gps.h
@@ -11,16 +11,16 @@ class gps{
     SFE_UBLOX_GNSS myGNSS;
     ErrorHandler* _errHand;
     bool GPSBegin();
-    int32_t latitude;
-    int32_t getLatitude();
-    int32_t longitude;
-    int32_t getLongitude();
-    int32_t altitude;
-    int32_t getAltitude();
+    float latitude;
+    float getLatitude();
+    float longitude;
+    float getLongitude();
+    long altitude;
+    long getAltitude();
     uint8_t SIV;
     uint8_t getSIV();
     private:
     bool working = false;
-    long lastTime = 0;
+    float lastTime = 0;
     TwoWire* _I2C;
 };
